parse, lex: Decompose AST extraction bottleneck and forward-declare dump types
- Enhance toolchain/parse/node_kind.def with category x-macros
  (CARBON_PARSE_NODE_KIND_DECLARATION, CARBON_PARSE_NODE_KIND_EXPRESSION,
  CARBON_PARSE_NODE_KIND_PATTERN, CARBON_PARSE_NODE_KIND_STATEMENT) that
  default to CARBON_PARSE_NODE_KIND, keeping node_kind.def as the single
  source of truth without manual expansion divergence.
- Decompose the monolithic toolchain/parse/extract.cpp translation unit
  into separate compilation units for declarations, expressions,
  patterns, and statements by expanding respective category x-macros.
- Forward-declare Tree and TokenizedBuffer in parse and lex dump headers
  to reduce header inclusion depth.
